+++ powercap/raplcap-powercap.c
@@ -23,8 +23,19 @@
 #define ZONE_NAME_MAX_SIZE 64
 #define ZONE_NAME_PREFIX_PACKAGE "package-"
 
+typedef struct raplcap_powercap_parent {
+  powercap_intel_rapl_parent p;
+  raplcap_zone type;
+  int has_pkg;
+  uint32_t pkg;
+  int has_die;
+  uint32_t die;
+} raplcap_powercap_parent;
+
 typedef struct raplcap_powercap {
-  powercap_intel_rapl_parent* parent_zones;
+  raplcap_powercap_parent* parent_zones;
+  raplcap_powercap_parent** pkg_zones;
+  raplcap_powercap_parent** psys_zones;
   uint32_t n_parent_zones;
   uint32_t n_pkg;
   // currently only support homogeneous die count per package
@@ -35,6 +46,7 @@ static raplcap rc_default;
 
 static powercap_intel_rapl_parent* get_parent_zone(const raplcap* rc, uint32_t pkg, uint32_t die, raplcap_zone zone) {
   raplcap_powercap* state;
+  raplcap_powercap_parent* p = NULL;
   if (rc == NULL) {
     rc = &rc_default;
   }
@@ -58,32 +70,48 @@ static powercap_intel_rapl_parent* get_parent_zone(const raplcap* rc, uint32_t p
     errno = EINVAL;
     return NULL;
   }
-  // PSYS is a special case - always stored at the end of the parent_zones array, with index > n_pkg
-  // If PSYS is requested and supported, it doesn't matter what pkg or die the caller actually specified
-  if (zone == RAPLCAP_ZONE_PSYS &&
-      powercap_intel_rapl_is_zone_supported(&state->parent_zones[state->n_parent_zones - 1], zone)) {
-    return &state->parent_zones[state->n_parent_zones - 1];
+  if (zone == RAPLCAP_ZONE_PSYS) {
+    // powercap control type doesn't specify die values for PSYS zones, so we assume die must be 0
+    if (die == 0) {
+      p = state->psys_zones[pkg];
+    }
+    // --- begin deprecation block
+    if (p == NULL) {
+      if ((p = state->psys_zones[pkg]) != NULL) {
+        raplcap_log(WARN, "Ignoring die value > 0 for PSYS zone at pkg=%"PRIu32".\nThis behavior is deprecated - "
+                    "in the future, an error will be returned if the zone is not found for the specified pkg/die.\n",
+                    pkg);
+      }
+      if (p == NULL && pkg > 0 && (p = state->psys_zones[0]) != NULL) {
+        raplcap_log(WARN, "Falling back on PSYS zone at pkg=0, die=0.\nThis behavior is deprecated - "
+                    "in the future, an error will be returned if the zone is not found for the specified pkg/die.\n");
+      }
+    }
+    // --- end deprecation block
+    // if p is still NULL, fall through and later code will (correctly) fail to find PSYS within the regular parent zone
   }
-  assert(((pkg * state->n_die) + die) < state->n_parent_zones);
-  return &state->parent_zones[(pkg * state->n_die) + die];
+  if (p == NULL) {
+    p = state->pkg_zones[(pkg * state->n_die) + die];
+  }
+  if (p == NULL) {
+    // the requested package/die was in range, but the zone was not detected in sysfs
+    errno = ENODEV;
+    return NULL;
+  }
+  return &p->p;
 }
 
-static uint32_t count_parent_zones(void) {
-  uint32_t n = powercap_intel_rapl_get_num_instances();
-  raplcap_log(DEBUG, "count_parent_zones: n=%"PRIu32"\n", n);
-  return n;
-}
-
-static int count_pkg_die(uint32_t* pkg_die_mask, uint32_t n_parent_zones, uint32_t* max_pkg_id) {
+static int get_topology(uint32_t *n_parent_zones, uint32_t* n_pkg, uint32_t* n_die) {
   char name[ZONE_NAME_MAX_SIZE];
   char* endptr;
   char* endptr2;
-  uint32_t i;
   uint32_t pkg;
   uint32_t die;
+  uint32_t max_pkg_id = 0;
+  uint32_t max_die_id = 0;
   // package and die IDs can appear in any order
-  for (i = 0; i < n_parent_zones; i++) {
-    if (powercap_sysfs_zone_get_name(CONTROL_TYPE, &i, 1, name, sizeof(name)) < 0) {
+  for (*n_parent_zones = 0; !powercap_sysfs_zone_exists(CONTROL_TYPE, n_parent_zones, 1); (*n_parent_zones)++) {
+    if (powercap_sysfs_zone_get_name(CONTROL_TYPE, n_parent_zones, 1, name, sizeof(name)) < 0) {
       raplcap_perror(ERROR, "powercap_sysfs_zone_get_name");
       return -1;
     }
@@ -98,16 +126,8 @@ static int count_pkg_die(uint32_t* pkg_die_mask, uint32_t n_parent_zones, uint32
       errno = EINVAL;
       return -1;
     }
-    if (pkg >= n_parent_zones) {
-      // something is weird with the kernel zone naming if this happens - packages w/ smaller IDs are missing
-      // NOTE: The presence of a PSYS zone should reduce the upper bound, even though the array index is still legal
-      //       Package IDs are verified later
-      raplcap_log(ERROR, "Package %"PRIu32" not in range [0, %"PRIu32")\n", pkg, n_parent_zones);
-      errno = EINVAL;
-      return -1;
-    }
-    if (*max_pkg_id < pkg) {
-      *max_pkg_id = pkg;
+    if (max_pkg_id < pkg) {
+      max_pkg_id = pkg;
     }
     if (*endptr == '\0') {
       // the string format is package-X
@@ -116,185 +136,138 @@ static int count_pkg_die(uint32_t* pkg_die_mask, uint32_t n_parent_zones, uint32
       // the string format is (presumably) package-X-die-Y
       die = strtoul(endptr + 1, &endptr2, 0);
       if (!die && (endptr + 1) == endptr2) {
-        //. failed to get die - something unexpected in name format
+        // failed to get die - something unexpected in name format
         raplcap_log(ERROR, "Failed to parse die from zone name: %s\n", name);
         errno = EINVAL;
         return -1;
+      }
+      if (max_die_id < die) {
+        max_die_id = die;
       }
     } else {
       raplcap_log(ERROR, "Unsupported zone name format: %s\n", name);
       errno = EINVAL;
       return -1;
     }
-    // supports detecting duplicate die entries up to die ID 32
-    if (pkg_die_mask[pkg] & (1 << die)) {
-      raplcap_log(ERROR, "Duplicate package/die detected, pkg=%"PRIu32", die=%"PRIu32"\n", pkg, die);
-      errno = EINVAL;
-      return -1;
-    }
-    pkg_die_mask[pkg] |= (1 << die);
   }
+  if (*n_parent_zones == 0) {
+    errno = ENODEV;
+    return -1;
+  }
+  *n_pkg = max_pkg_id + 1;
+  *n_die = max_die_id + 1;
+  raplcap_log(DEBUG, "get_topology: n_parent_zones=%"PRIu32", n_pkg=%"PRIu32", n_die=%"PRIu32"\n",
+              *n_parent_zones, *n_pkg, *n_die);
   return 0;
 }
 
-static int parse_pkg_die(const uint32_t* pkg_die_mask, uint32_t n_parent_zones, uint32_t max_pkg_id, uint32_t* n_pkgs, uint32_t* n_die) {
-  uint32_t i;
-  uint32_t n_pkg_die;
-  for (i = 0; i < n_parent_zones; i++) {
-    if (!pkg_die_mask[i]) {
-      // not a PACKAGE zone (e.g., could be PSYS)
-      continue;
-    }
-    (*n_pkgs)++;
-    // count die
-    n_pkg_die = __builtin_popcount(pkg_die_mask[i]);
-    // only the lowest bits should be set, otherwise some die are missing
-    if ((pkg_die_mask[i] >> n_pkg_die) != 0) {
-      raplcap_log(ERROR, "Unexpected package/die mapping - possibly missing package die\n");
-      errno = EINVAL;
-      return -1;
-    }
-    if (*n_die == 0) {
-      *n_die = n_pkg_die;
-    } else if (n_pkg_die != *n_die) {
-      // currently only support uniform package die counts, so verify that every package has matching die count
-      raplcap_log(ERROR, "Unsupported heterogeneity in package/die configuration\n");
-      errno = EINVAL;
-      return -1;
-    }
-  }
-  // verify that packages aren't missing from range [0, n_pkgs)
-  if (max_pkg_id != *n_pkgs - 1) {
-    raplcap_log(ERROR, "Unexpected package/die mapping - possibly missing package\n");
-    errno = EINVAL;
-    return -1;
-  }
-  return 0;
-}
-
-static int get_topology(uint32_t n_parent_zones, uint32_t* n_pkgs, uint32_t* n_die) {
-  uint32_t* pkg_die_mask;
-  uint32_t max_pkg_id = 0;
-  int ret = 0;
-  if (!n_parent_zones) {
-    errno = EINVAL;
-    return -1;
-  }
-  if (!(pkg_die_mask = calloc(n_parent_zones, sizeof(uint32_t)))) {
-    return -1;
-  }
-  *n_pkgs = 0;
-  *n_die = 0;
-  // TODO: Enforce that pkg IDs fill the range [0, n_pkgs), and die IDs fill the range [0, n_die) for each package
-  //       This is required to correctly access zones for "pkg" and "die" params used to compute array indexes
-  if (!(ret = count_pkg_die(pkg_die_mask, n_parent_zones, &max_pkg_id))) {
-    ret = parse_pkg_die(pkg_die_mask, n_parent_zones, max_pkg_id, n_pkgs, n_die);
-  }
-  raplcap_log(DEBUG, "get_topology: packages=%"PRIu32", die=%"PRIu32"\n", *n_pkgs, *n_die);
-  free(pkg_die_mask);
-  return ret;
-}
-
-// compare strings that may contain substrings of natural numbers (values >= 0)
-// format is expected to be: "package-%d" or "package-%d-die-%d"
-static int strcmp_nat_lu(const char* l, const char* r) {
-  unsigned long l_val;
-  unsigned long r_val;
-  long diff;
+static int parse_parent_zone_topology(raplcap_powercap_parent* rp, uint32_t id) {
+  char name[ZONE_NAME_MAX_SIZE] = { 0 };
+  const char* ptr = name;
   char* endptr;
-  int is_num_block = 0;
-  while (*l && *r) {
-    if (is_num_block) {
-      // parse numeric values
-      l_val = strtoul(l, &endptr, 0);
-      l = endptr;
-      r_val = strtoul(r, &endptr, 0);
-      r = endptr;
-      diff = l_val - r_val;
-      if (diff > 0) {
-        return 1;
-      }
-      if (diff < 0) {
+  // first determine zone type
+  if (powercap_intel_rapl_is_zone_supported(&rp->p, RAPLCAP_ZONE_PACKAGE)) {
+    rp->type = RAPLCAP_ZONE_PACKAGE;
+  } else if (powercap_intel_rapl_is_zone_supported(&rp->p, RAPLCAP_ZONE_PSYS)) {
+    rp->type = RAPLCAP_ZONE_PSYS;
+  } else {
+    raplcap_log(ERROR, "Unexpected type for parent zone id=%"PRIu32"\n", id);
+    errno = ENOTSUP;
+    return -1;
+  }
+  raplcap_log(DEBUG, "parse_parent_zone_topology: id=%"PRIu32", type=%u\n", id, rp->type);
+  // now parse name for pkg and die, if available
+  // name format is expected to be: "package-%d", "package-%d-die-%d", "psys", or "psys-%d"
+  if (powercap_intel_rapl_get_name(&rp->p, rp->type, name, sizeof(name)) < 0) {
+    raplcap_perror(ERROR, "powercap_intel_rapl_get_name");
+    return -1;
+  }
+  while (*ptr) {
+    if (isdigit(*ptr)) {
+      if (!rp->has_pkg) {
+        endptr = NULL;
+        rp->pkg = strtoul(ptr, &endptr, 0);
+        if (!rp->pkg && endptr == ptr) {
+          raplcap_log(ERROR, "Failed to parse package from zone name: %s\n", name);
+          errno = ENOTSUP;
+          return -1;
+        }
+        rp->has_pkg = 1;
+        raplcap_log(DEBUG, "parse_parent_zone_topology: id=%"PRIu32", pkg=%"PRIu32"\n", id, rp->pkg);
+      } else if (!rp->has_die) {
+        endptr = NULL;
+        rp->die = strtoul(ptr, &endptr, 0);
+        if (!rp->die && endptr == ptr) {
+          raplcap_log(ERROR, "Failed to parse die from zone name: %s\n", name);
+          errno = ENOTSUP;
+          return -1;
+        }
+        rp->has_die = 1;
+        raplcap_log(DEBUG, "parse_parent_zone_topology: id=%"PRIu32", die=%"PRIu32"\n", id, rp->die);
+      } else {
+        raplcap_log(ERROR, "Unsupported name format for parent zone id=%"PRIu32": %s\n", id, name);
+        errno = ENOTSUP;
         return -1;
       }
-      is_num_block = 0;
+      ptr = endptr;
     } else {
-      // compare non-numerically until both strings are digits at same index
-      for (; *l && *r; l++, r++) {
-        if (isdigit(*l) && isdigit(*r)) {
-          is_num_block = 1;
-          break;
-        }
-        if (isdigit(*l)) {
-          return -1;
-        }
-        if (isdigit(*r)) {
-          return 1;
-        }
-        diff = *l - *r;
-        if (diff > 0) {
-          return 1;
-        }
-        if (diff < 0) {
-          return -1;
-        }
-      }
+      ptr++;
     }
   }
-  return *r ? -1 : (*l ? 1 : 0);
+  return 0;
 }
 
-static int sort_parent_zones(const void* a, const void* b) {
-  char name_a[ZONE_NAME_MAX_SIZE] = { 0 };
-  char name_b[ZONE_NAME_MAX_SIZE] = { 0 };
-  const powercap_intel_rapl_parent* zone_a = (const powercap_intel_rapl_parent*) a;
-  const powercap_intel_rapl_parent* zone_b = (const powercap_intel_rapl_parent*) b;
-  int ret = 0;
-  // first check if either zone is PSYS, which must be placed at the end of the sorted array (after PACKAGE zones)
-  // in the current powercap design, PSYS zones exist as their own parent zones, without child zones
-  if (powercap_intel_rapl_is_zone_supported(zone_a, RAPLCAP_ZONE_PSYS)) {
-    return 1; // a >= b
+// rp is expected to be zero-initialized
+static int raplcap_powercap_parent_init(raplcap_powercap_parent* rp, uint32_t id, int ro) {
+  int err_save;
+  if (powercap_intel_rapl_init(id, &rp->p, ro)) {
+    raplcap_perror(ERROR, "powercap_intel_rapl_init");
+    return -1;
   }
-  if (powercap_intel_rapl_is_zone_supported(zone_b, RAPLCAP_ZONE_PSYS)) {
-    return -1; // a < b
+  if (parse_parent_zone_topology(rp, id)) {
+    err_save = errno;
+    powercap_intel_rapl_destroy(&rp->p);
+    errno = err_save;
+    return -1;
   }
-  // now assume PACKAGE zones and sort by name
-  if (powercap_intel_rapl_get_name(zone_a, RAPLCAP_ZONE_PACKAGE, name_a, sizeof(name_a)) >= 0 &&
-      powercap_intel_rapl_get_name(zone_b, RAPLCAP_ZONE_PACKAGE, name_b, sizeof(name_b)) >= 0) {
-    // assumes names are in the form "package-X" or "package-X-die-Y"
-    if ((ret = strcmp_nat_lu(name_a, name_b)) > 0) {
-      raplcap_log(DEBUG, "sort_parent_zones: Zones are out of order\n");
-    }
-  } else {
-    raplcap_perror(ERROR, "powercap_intel_rapl_get_name");
-  }
-  return ret;
+  return 0;
 }
 
 int raplcap_init(raplcap* rc) {
   raplcap_powercap* state;
-  uint32_t n_parent_zones;
+  uint32_t n_parent_zones = 0;
   uint32_t n_pkg;
   uint32_t n_die;
   uint32_t i;
+  uint32_t pkg;
+  uint32_t die;
   int err_save;
   const char* env_ro = getenv(ENV_RAPLCAP_READ_ONLY);
   int ro = env_ro == NULL ? 0 : atoi(env_ro);
   if (rc == NULL) {
     rc = &rc_default;
   }
-  // get the number of packages
-  if ((n_parent_zones = count_parent_zones()) == 0) {
-    raplcap_perror(ERROR, "No RAPL zones found");
-    return -1;
-  }
-  if (get_topology(n_parent_zones, &n_pkg, &n_die) < 0) {
+  if (get_topology(&n_parent_zones, &n_pkg, &n_die) < 0) {
+    if (n_parent_zones == 0) {
+      raplcap_perror(ERROR, "No RAPL zones found");
+    }
     return -1;
   }
   if ((state = malloc(sizeof(raplcap_powercap))) == NULL) {
     return -1;
   }
-  if ((state->parent_zones = malloc(n_parent_zones * sizeof(powercap_intel_rapl_parent))) == NULL) {
+  if ((state->parent_zones = calloc(n_parent_zones, sizeof(*state->parent_zones))) == NULL) {
+    free(state);
+    return -1;
+  }
+  if ((state->pkg_zones = calloc(n_pkg * n_die, sizeof(*state->pkg_zones))) == NULL) {
+    free(state->parent_zones);
+    free(state);
+    return -1;
+  }
+  if ((state->psys_zones = calloc(n_pkg, sizeof(*state->psys_zones))) == NULL) {
+    free(state->pkg_zones);
+    free(state->parent_zones);
     free(state);
     return -1;
   }
@@ -303,8 +276,7 @@ int raplcap_init(raplcap* rc) {
   state->n_die = n_die;
   rc->state = state;
   for (i = 0; i < state->n_parent_zones; i++) {
-    if (powercap_intel_rapl_init(i, &state->parent_zones[i], ro)) {
-      raplcap_perror(ERROR, "powercap_intel_rapl_init");
+    if (raplcap_powercap_parent_init(&state->parent_zones[i], i, ro)) {
       err_save = errno;
       state->n_parent_zones = i; // so as not to cleanup uninitialized zones
       raplcap_destroy(rc);
@@ -312,16 +284,37 @@ int raplcap_init(raplcap* rc) {
       return -1;
     }
   }
-  // it's been observed that packages in sysfs may be numbered out of order; we must sort by name
-  // if there are PSYS zones, we sort them to the end of the array
-  errno = 0;
-  qsort(state->parent_zones, state->n_parent_zones, sizeof(powercap_intel_rapl_parent), sort_parent_zones);
-  if (errno) {
-    raplcap_log(ERROR, "Failed to sort packages by name\n");
-    err_save = errno;
-    raplcap_destroy(rc);
-    errno = err_save;
-    return -1;
+  // Parent zones in sysfs may be out of order - index by type, package, and die
+  for (i = 0; i < state->n_parent_zones; i++) {
+    pkg = state->parent_zones[i].pkg;
+    die = state->parent_zones[i].die;
+    if (pkg >= n_pkg || die >= n_die) {
+      // this should only arise if sysfs has changed since we initially parsed topology - unlikely, but possible...
+      raplcap_log(ERROR, "Package or die out of range for parent zone id=%"PRIu32"\n", i);
+      err_save = errno;
+      raplcap_destroy(rc);
+      errno = err_save;
+      return -1;
+    }
+    switch (state->parent_zones[i].type) {
+      case RAPLCAP_ZONE_PACKAGE:
+        if (state->pkg_zones[(pkg * n_die) + die] == NULL) {
+          state->pkg_zones[(pkg * n_die) + die] = &state->parent_zones[i];
+        } else {
+          raplcap_log(WARN, "Ignoring duplicate package entry at parent zone id=%"PRIu32"\n", i);
+        }
+        break;
+      case RAPLCAP_ZONE_PSYS:
+        if (state->psys_zones[pkg] == NULL) {
+          state->psys_zones[pkg] = &state->parent_zones[i];
+        } else {
+          raplcap_log(WARN, "Ignoring duplicate psys entry at parent zone id=%"PRIu32"\n", i);
+        }
+        break;
+      default:
+        raplcap_log(WARN, "Ignoring unknown type at parent zone id=%"PRIu32"\n", i);
+        break;
+    }
   }
   rc->nsockets = n_pkg;
   raplcap_log(DEBUG, "raplcap_init: Initialized\n");
@@ -338,11 +331,13 @@ int raplcap_destroy(raplcap* rc) {
   if ((state = (raplcap_powercap*) rc->state) != NULL) {
     for (i = 0; i < state->n_parent_zones; i++) {
       raplcap_log(DEBUG, "raplcap_destroy: zone=%"PRIu32"\n", i);
-      if (powercap_intel_rapl_destroy(&state->parent_zones[i])) {
+      if (powercap_intel_rapl_destroy(&state->parent_zones[i].p)) {
         raplcap_perror(WARN, "powercap_intel_rapl_destroy");
         err_save = errno;
       }
     }
+    free(state->psys_zones);
+    free(state->pkg_zones);
     free(state->parent_zones);
     free(state);
     rc->state = NULL;
@@ -353,16 +348,9 @@ int raplcap_destroy(raplcap* rc) {
   return err_save ? -1 : 0;
 }
 
-static int get_topology_uninit(uint32_t* n_pkg, uint32_t* n_die) {
-  uint32_t n_parent_zones;
-  if ((n_parent_zones = count_parent_zones()) == 0) {
-    return -1;
-  }
-  return get_topology(n_parent_zones, n_pkg, n_die);
-}
-
 uint32_t raplcap_get_num_packages(const raplcap* rc) {
   const raplcap_powercap* state;
+  uint32_t n_parent_zones;
   uint32_t n_pkg = 0;
   uint32_t n_die;
   if (rc == NULL) {
@@ -371,13 +359,14 @@ uint32_t raplcap_get_num_packages(const raplcap* rc) {
   if ((state = (raplcap_powercap*) rc->state) != NULL) {
     return state->n_pkg;
   }
-  return get_topology_uninit(&n_pkg, &n_die) ? 0 : n_pkg;
+  return get_topology(&n_parent_zones, &n_pkg, &n_die) ? 0 : n_pkg;
 }
 
 uint32_t raplcap_get_num_die(const raplcap* rc, uint32_t pkg) {
   const raplcap_powercap* state;
+  uint32_t n_parent_zones;
   uint32_t n_pkg;
-  uint32_t n_die;
+  uint32_t n_die = 0;
   if (rc == NULL) {
     rc = &rc_default;
   }
@@ -389,7 +378,7 @@ uint32_t raplcap_get_num_die(const raplcap* rc, uint32_t pkg) {
     }
     return state->n_die;
   }
-  if (get_topology_uninit(&n_pkg, &n_die)) {
+  if (get_topology(&n_parent_zones, &n_pkg, &n_die)) {
     return 0;
   }
   if (pkg >= n_pkg) {